Disable stats derivation for CLogicalFullOuterJoin
Instead, use the stats derived from an equivalent UNION ALL expression.  To
maintain the ability to enable/disable the ExfExpandFullOuterJoin, this commit
introduces a new traceflag.

Instead of enabling or disabling the transform directly, it skips the
implementation of the UnionAll, if the trace flag is disabled and the UnionAll
op was generated by ExfExpandFullOuterJoin.  This way, even if expansion is
disabled by the trace flag, the xform is still triggered, so that the UnionAll
operator generated can be used to compute statistics for the logical merge join
group (see CLogicalFullOuterJoin::Esp()).